added update page

+++ controller/admin/PagesController.php
@@ -12,11 +12,32 @@ class AdminPagesController extends BaseController {
         authenticateForward();
 
         if ($this -> request -> getAction() == 'Success') {
-            $this -> response -> addNotification($this -> request -> getName().' is created.');
+            if (strpos($_SERVER['HTTP_REFERER'], '/admin/page/add') !== false) {
+                $this -> response -> addNotification($this -> request -> getName().' is created.');
+            } elseif (strpos($_SERVER['HTTP_REFERER'], '/admin/page/edit') !== false) {
+                $this -> response -> addNotification($this -> request -> getName().' is updated.');
+            }
         } elseif ($this -> request -> getAction() == 'Delete') {
-            $count = PageDao::delete($this -> request -> getPageId());
-            if ($count > 0) {
-                $this -> response -> addNotification('Delete successful.');
+            if ($this -> request -> valid()) {
+                $count = PageDao::delete($this -> request -> getPageId());
+                if ($count > 0) {
+                    $this -> response -> addNotification('Delete successful.');
+
+                    foreach ($this -> request -> getPageId() as $page_id) {
+                        $page_model = new PageModel();
+                        $page_model -> setId($page_id);
+                        $page_model -> setPageTypeId(1);
+
+                        $filename = PageUtil::generateFilename($page_model);
+                        PageUtil::deleteHtml($filename);
+                        PageUtil::deleteCss($filename);
+                        PageUtil::deleteJs($filename);
+                        PageUtil::deleteController($filename);
+                        
+                    }
+                }
+            } else {
+                $this -> response -> setError($this -> request -> getErrors());
             }
         }
